requirements.txt
herbie-data
pandas
pyarrow
numpy
contourpy
scipy
simplification